operator test
